refactor(backend-http): move modules for architecture-consistency
