+++ packages/backend-http/src/http/application/builders/ResponseBuilder.ts
@@ -1,7 +1,7 @@
 import { Builder } from '@one-game-js/backend-common';
 
-import { Response } from '../../models/application/Response';
-import { ResponseWithBody } from '../../models/application/ResponseWithBody';
+import { Response } from '../models/Response';
+import { ResponseWithBody } from '../models/ResponseWithBody';
 
 export abstract class ResponseBuilder<TParams extends unknown[]>
   implements Builder<Response | ResponseWithBody<unknown>, TParams>
