+++ packages/backend-http/src/http/application/controllers/SingleEntityHttpRequestController.ts
@@ -1,9 +1,9 @@
 import { Builder, Handler } from '@one-game-js/backend-common';
 
-import { Request } from '../../models/application/Request';
-import { RequestWithBody } from '../../models/application/RequestWithBody';
-import { Response } from '../../models/application/Response';
-import { ResponseWithBody } from '../../models/application/ResponseWithBody';
+import { Request } from '../models/Request';
+import { RequestWithBody } from '../models/RequestWithBody';
+import { Response } from '../models/Response';
+import { ResponseWithBody } from '../models/ResponseWithBody';
 
 export abstract class SingleEntityHttpRequestController<
   TRequest extends Request | RequestWithBody = Request | RequestWithBody,
