+++ packages/backend-http/src/index.ts
@@ -1,11 +1,11 @@
-import { ErrorV1ResponseFromErrorBuilder } from './http/builders/application/ErrorV1ResponseFromErrorBuilder';
-import { SingleEntityGetResponseBuilder } from './http/builders/application/SingleEntityGetResponseBuilder';
-import { SingleEntityPostResponseBuilder } from './http/builders/application/SingleEntityPostResponseBuilder';
-import { SingleEntityHttpRequestController } from './http/controllers/application/SingleEntityHttpRequestController';
-import { Request } from './http/models/application/Request';
-import { RequestWithBody } from './http/models/application/RequestWithBody';
-import { Response } from './http/models/application/Response';
-import { ResponseWithBody } from './http/models/application/ResponseWithBody';
+import { ErrorV1ResponseFromErrorBuilder } from './http/application/builders/ErrorV1ResponseFromErrorBuilder';
+import { SingleEntityGetResponseBuilder } from './http/application/builders/SingleEntityGetResponseBuilder';
+import { SingleEntityPostResponseBuilder } from './http/application/builders/SingleEntityPostResponseBuilder';
+import { SingleEntityHttpRequestController } from './http/application/controllers/SingleEntityHttpRequestController';
+import { Request } from './http/application/models/Request';
+import { RequestWithBody } from './http/application/models/RequestWithBody';
+import { Response } from './http/application/models/Response';
+import { ResponseWithBody } from './http/application/models/ResponseWithBody';
 
 export type { Request, RequestWithBody, Response, ResponseWithBody };
